Added gson. Cleaned up.
