+++ app/src/main/java/com/iit/dashboard2022/MainActivity.java
@@ -17,15 +17,12 @@ import com.iit.dashboard2022.page.LiveData;
 import com.iit.dashboard2022.page.Logs;
 import com.iit.dashboard2022.page.PageManager;
 import com.iit.dashboard2022.page.Pager;
-import com.iit.dashboard2022.telemetry.TelemetryHandler;
 import com.iit.dashboard2022.ui.SidePanel;
 import com.iit.dashboard2022.ui.widget.SettingsButton;
 import com.iit.dashboard2022.ui.widget.WidgetUpdater;
 import com.iit.dashboard2022.ui.widget.console.ConsoleWidget;
 import com.iit.dashboard2022.util.HawkUtil;
 import com.iit.dashboard2022.util.Toaster;
-
-import java.net.InetSocketAddress;
 
 public class MainActivity extends AppCompatActivity {
