+++ app/src/main/java/com/iit/dashboard2022/util/HawkUtil.java
@@ -1,5 +1,8 @@
 package com.iit.dashboard2022.util;
 
+import android.content.Context;
+import android.net.ConnectivityManager;
+import android.net.NetworkCapabilities;
 import android.os.Build;
 import android.view.View;
 import android.view.Window;
@@ -29,4 +32,20 @@ public class HawkUtil {
                                                         | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION);
         }
     }
+
+    /**
+     * Checks if the host device has a network connection
+     *
+     * @param context {@link Context}
+     * @return True if host has connection, false if not
+     */
+    public static boolean checkInternetConnection(Context context) {
+        ConnectivityManager cm = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
+        if (cm != null) {
+            NetworkCapabilities capabilities = cm.getNetworkCapabilities(cm.getActiveNetwork());
+            return capabilities != null && (capabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) || capabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) || capabilities.hasTransport(NetworkCapabilities.TRANSPORT_VPN));
+        }
+        return false;
+    }
+
 }
